updated commands

+++ Blog.Service.BlogApi.Application/Features/Comments/Commands/DeleteComment/DeleteCommentHandler.cs
@@ -1,4 +1,5 @@
-﻿using Blog.Service.BlogApi.Domain.Repositories;
+﻿using Blog.Service.BlogApi.Domain.Comments;
+using Blog.Service.BlogApi.Domain.Repositories;
 using MediatR;
 using System.Threading;
 using System.Threading.Tasks;
@@ -16,10 +17,18 @@ namespace Blog.Service.BlogApi.Application.Features.Comments.Commands.DeleteComm
 
         public async Task<bool> Handle(DeleteCommentCommand request, CancellationToken cancellationToken)
         {
+            Comment entity = _blogUnitOfWork.CommentReadOnlyRepository.Get(request.PostId, request.Id);
+
+            if (entity == null) throw new Exceptions.ApplicationException("No comment is found to delete");
+
+            if (!entity.UserId.Equals(request.UserId))
+            {
+                throw new Exceptions.ApplicationException("Unauthorized to delete comment"); //need new exeption model class
+            }
+
             var isSucceed = _blogUnitOfWork.CommentCommandRepository.Delete(request.PostId ,request.Id);
 
             if (!isSucceed) throw new Exceptions.ApplicationException("Failed to update comment");
-
 
             return true;
         }
